committing work before major change - about to switch all variables to double rather than float.  may fail at this endeavor, so this is the one where everything is working before the change.
git-svn-id: file:///home/byron/Desktop/svn/archival/RealEstateMarketAnalysis/branches/version-0_75@181 6aeec067-7811-e111-a833-485b39788f86

+++ src/com/byronkatz/reap/calculations/ModifiedInternalRateOfReturn.java
@@ -8,14 +8,14 @@ import com.byronkatz.reap.general.ValueEnum;
 public class ModifiedInternalRateOfReturn {
 
   private Vector<Float> cashFlowVector;
-  private Float yearlyInterestRate;
+  private Double yearlyInterestRate;
   private Float yearlyRequiredRateOfReturn;
   private Float mirrValueAccumulator;
 
   private DataController dataController;
 
   public ModifiedInternalRateOfReturn(DataController dataController,
-      Float yearlyInterestRate, 
+      Double yearlyInterestRate, 
       Float yearlyRequiredRateOfReturn) {
     this.dataController = dataController;
     mirrValueAccumulator = 0.0f;
